requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
SQLAlchemy-serializer==1.4.1
python-dotenv==1.0.0